Add SortedByItem impl for affine iterators
Affine iterators have 0 or 1 item, they are trivially sorted.
In the rust std, it is surprisingly easy to track if an iterator has 0
or 1 item! So let's do it.

In addition, this allows accepting as sorted a `iter::FlatMap`,
`iter::Map` and `iter::Flatten` if it was called on an affine iterator.

My use case for this is iterating over a `Option<&BTreeMap<…>>`.